Move function getBibleBookStatsFromVerseList to from AppController to new verseListController
